Last commit, pretty much done with Boat commands
This commit is good to have, as the game is now really playable. 
This will be the last commit for this card, and the branch and trello card will be closed.

+++ worldofzuul/DisplayInterfaces/DisplayAddedGold.java
@@ -1,6 +1,6 @@
 package worldofzuul.DisplayInterfaces;
 
 public interface DisplayAddedGold {
-    void displayNewGold(int amount);
+    void displayNewGold(double amount);
     void displayGold(double amount);
 }
